Added CustomAnalyzer and Utilities
Added the CustomAnalyzer class, which can take stemming algorithms as inputs.

Added the "CommonTermsLowFreqBoostQuery" class,  a subclass of the "CommonTermsQuery". Can find rare terms in a query and automatically apply boosts to them.

The "MultiQueryTermExctractor" class is used to parse Multi-field queries in to the CommonTermsLowFreqBoostQuery, but it may not be needed.